move challenges to a custom component
